+++ Team32/Code32/src/spa/src/QP/QueryGraph.h
@@ -14,7 +14,7 @@ public:
 	explicit QueryGraph(const DeclarationList& declarations);
 	void setEdges(const ClauseList& clause_list);
 	unordered_map<string, Node> getNodes();
-	unordered_map<string, int> getSynonymsInGroup(const string& selected_synonym);
+	unordered_map<string, size_t> getSynonymsInGroup(const string& selected_synonym);
 
 private:
 	unordered_map<string, Node> nodes;
