+++ Team32/Code32/src/spa/src/QP/QueryResult.h
@@ -28,8 +28,8 @@ private:
 	void joinWithDifferentSynonym(QueryResult& query_result);
 	void joinWithSameSynonym(QueryResult& query_result);
 	static bool contains(unordered_map<string, vector<string>>& table, const unordered_map<string, string>& row);
-	static bool isRowMatch(const unordered_map<string, string>& row, unordered_map<string, vector<string>>& table, int row_number);
-	static void removeRow(unordered_map<string, vector<string>>& table, int row_number);
+	static bool isRowMatch(const unordered_map<string, string>& row, unordered_map<string, vector<string>>& table, size_t row_number);
+	static void removeRow(unordered_map<string, vector<string>>& table, size_t row_number);
 	unordered_map<string, vector<string>> getSubTableWithRow(const unordered_map<string, string>& row);
 };
